backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
xgboost==2.1.1
lightgbm==4.5.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2
groq==0.11.0
supabase==2.7.4
aiohttp==3.10.5
beautifulsoup4==4.12.3
requests==2.32.3
python-multipart==0.0.12
